Correctly introspect double-precision float in postgres.
Fixes #1988

+++ playhouse/reflection.py
@@ -224,7 +224,7 @@ class PostgresqlMetadata(Metadata):
         23: IntegerField,
         25: TextField,
         700: FloatField,
-        701: FloatField,
+        701: DoubleField,
         1042: CharField, # blank-padded CHAR
         1043: CharField,
         1082: DateField,
